Move metrics to subdirectory
