--- llarp/util/metrics.cpp
@@ -1,1 +1,0 @@
-#include <util/metrics.hpp>
