+++ llarp/util/metrics/core.cpp
@@ -1,4 +1,4 @@
-#include <util/metrics_core.hpp>
+#include <util/metrics/core.hpp>
 
 #include <iostream>
