+++ llarp/util/metrics/json_publisher.cpp
@@ -1,4 +1,4 @@
-#include <metrics/json_publisher.hpp>
+#include <util/metrics/json_publisher.hpp>
 
 #include <fstream>
 #include <iomanip>
