+++ llarp/util/metrics/metrics.cpp
@@ -1,0 +1,1 @@
+#include <util/metrics/metrics.hpp>
