+++ llarp/util/metrics/metrics.hpp
@@ -1,8 +1,8 @@
 #ifndef LLARP_METRICS_HPP
 #define LLARP_METRICS_HPP
 
-#include <util/metrics_types.hpp>
-#include <util/metrics_core.hpp>
+#include <util/metrics/types.hpp>
+#include <util/metrics/core.hpp>
 #include <util/string_view.hpp>
 
 namespace llarp
