+++ llarp/util/metrics/stream_publisher.cpp
@@ -1,4 +1,4 @@
-#include <metrics/stream_publisher.hpp>
+#include <util/metrics/stream_publisher.hpp>
 
 #include <fstream>
 #include <iostream>
