+++ llarp/util/metrics/types.cpp
@@ -1,4 +1,4 @@
-#include <util/metrics_types.hpp>
+#include <util/metrics/types.hpp>
 
 #include <util/printer.hpp>
