+++ test/util/metrics/test_llarp_util_metrics_core.cpp
@@ -1,4 +1,4 @@
-#include <util/metrics_core.hpp>
+#include <util/metrics/core.hpp>
 
 #include <array>
 #include <thread>
