+++ test/util/metrics/test_llarp_util_metrics_types.cpp
@@ -1,4 +1,4 @@
-#include <util/metrics_types.hpp>
+#include <util/metrics/types.hpp>
 
 #include <array>
